requirements.txt
pymupdf4llm
slack_bolt
slack_sdk
python-dotenv
requests
anthropic
mammoth
ocrmypdf
markdownify
